pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mutcli"
version = "0.1.1"
description = "Mobile UI Testing CLI - Run YAML-based mobile tests anywhere"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Vladislav Karpman" }
]
keywords = ["mobile", "testing", "android", "ui", "automation", "scrcpy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "mysc>=0.5.0",           # MYScrcpy - scrcpy 3.x client
    "adbutils>=2.0.0",       # Device management
    "pillow>=10.0.0",        # Image processing
    "numpy>=1.24.0",         # Frame handling
    "av>=12.0.0",            # PyAV - video encoding
    "google-genai>=1.0.0",   # Gemini API client
    "pyyaml>=6.0",           # YAML parsing
    "typer>=0.12.0",         # CLI framework
    "rich>=13.0.0",          # Terminal output
    "python-dotenv>=1.0.0",  # Load .env files
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
mut = "mutcli.cli:app"

[project.urls]
Homepage = "https://github.com/vladkarpman/mut"
Documentation = "https://github.com/vladkarpman/mut#readme"
Repository = "https://github.com/vladkarpman/mut"
Issues = "https://github.com/vladkarpman/mut/issues"

[tool.hatch.build.targets.wheel]
packages = ["mutcli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
